商品削除機能修正

+++ app/controllers/items_controller.rb
@@ -36,6 +36,7 @@ class ItemsController < ApplicationController
   end
 
   def destroy
+    redirect_to item_path unless current_user.id == @item.user_id
     if @item.destroy
       redirect_to root_path
     else
